# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APP API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/business-feed/{businessId}": {
            "get": {
                "operationId": "App\\Http\\Controllers\\API\\v1\\BusinessFeedController::index",
                "parameters": [
                    {
                        "name": "business_id",
                        "in": "query",
                        "description": "ID of business",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of FeedResource"
                    }
                }
            }
        },
        "/api/v1/business-hours/{id}": {
            "put": {
                "operationId": "App\\Http\\Controllers\\API\\v1\\BusinessHoursController::updateOpenHours",
                "parameters": [
                    {
                        "name": "open_period_mins",
                        "in": "query",
                        "description": "Start Time",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "close_period_mins",
                        "in": "query",
                        "description": "end time",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "business_id",
                        "in": "query",
                        "description": "ID of business",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "wd_0",
                        "in": "query",
                        "description": "Sunday",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "wd_1",
                        "in": "query",
                        "description": "Monday",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "wd_2",
                        "in": "query",
                        "description": "Tueday",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "wd_3",
                        "in": "query",
                        "description": "Wednesday",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "wd_4",
                        "in": "query",
                        "description": "Thursday",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "wd_5",
                        "in": "query",
                        "description": "Friday",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "wd_6",
                        "in": "query",
                        "description": "Saturday",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "BusinessHoursResource"
                    }
                }
            }
        },
        "/api/v1/business-posts": {
            "get": {
                "operationId": "App\\Http\\Controllers\\API\\v1\\BusinessPostsController::index",
                "parameters": [
                    {
                        "name": "business_id",
                        "in": "query",
                        "description": "ID of business",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of Businesse Post"
                    }
                }
            },
            "put": {
                "operationId": "App\\Http\\Controllers\\API\\v1\\BusinessPostsController::update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "ID of business post",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "business_id",
                        "in": "query",
                        "description": "ID of business",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "text",
                        "in": "query",
                        "description": "Text of business post",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "expire_date",
                        "in": "query",
                        "description": "Date",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Business post updated"
                    },
                    "400": {
                        "description": "Invalid given data"
                    },
                    "404": {
                        "description": "Business post not found"
                    }
                }
            },
            "post": {
                "operationId": "App\\Http\\Controllers\\API\\v1\\BusinessPostsController::store",
                "parameters": [
                    {
                        "name": "business_id",
                        "in": "query",
                        "description": "ID of business",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "text",
                        "in": "query",
                        "description": "ID of user",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "photo",
                        "in": "query",
                        "description": "Image encoded in base64",
                        "schema": {
                            "type": "image"
                        }
                    },
                    {
                        "name": "expire_date",
                        "in": "query",
                        "description": "I",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "BusinessPostResource"
                    }
                }
            }
        },
...
    }
}